Convert ScoreWeight to integer
That's what it originally was

+++ dynamicbeat/pkg/check/check.go
@@ -12,11 +12,11 @@ type Check interface {
 }
 
 type Metadata struct {
-	ID          string  `json:"id"`
-	Name        string  `json:"name"`
-	Type        string  `json:"type"`
-	Group       string  `json:"group"`
-	ScoreWeight float64 `json:"score_weight"`
+	ID          string `json:"id"`
+	Name        string `json:"name"`
+	Type        string `json:"type"`
+	Group       string `json:"group"`
+	ScoreWeight int32  `json:"score_weight"`
 }
 
 type Config struct {
